requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dj-rest-auth==7.0.1
Django==5.1.5
django-allauth==65.3.1
django-cors-headers==4.6.0
django-filter==24.3
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
joblib==1.4.2
lxml==5.3.0
Markdown==3.7
nltk==3.9.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-http-client==3.3.7
regex==2024.11.6
requests==2.32.3
sendgrid==6.11.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
